queue-times-ingestion-cloud-function/requirements.txt
requests
google-cloud-storage
google-cloud-logging